Add optional noTranscriptClean to skip isoseq3
The workflow currently has transcript polishing with isoseq3 as the
final step. This only makes sense for RNA PacBio reads, while the rest
of the workflow can be use for both DNA and RNA.

By adding an option to disable the isoseq3 transcript polishing, this
workflow can also be used for DNA PacBio reads.

This resolves #issue-3