Rename to runIsoseq3Refine and disable by default
The option has been renamed to runIsoseq3Refine to be more explicit
about which step from isoseq3 is used, and is now disabled by default.
This has been done because this step is an additional step that only
makes sense when analysing data from RNA transcripts, while the other
steps in the pipeline are shared for both DNA and RNA reads.