Add sample names to workflow output
The sample names are determined by the content of the barcodes fasta
file, which is not accessible for the workflow. This change extracts the
sample names from the lima output files, and adds them to the output as
an array of strings. This way, workflows that incorporate
PacBio-subreads-processing can access the sample names directly